Move home tests to specific folder
